+++ browser/mainapp/src/util/Disposable.ts
@@ -35,6 +35,14 @@ export function isDisposable(obj: any) {
         && typeof obj[Symbol.dispose] == "function");
 }
 
+export function maybeDispose(obj: any) {
+    if (isDisposable(obj)) {
+        (obj as Disposable)[Symbol.dispose]();
+    }
+    else if (isIDisposable(obj)) {
+        (obj as IDisposable).dispose();
+    }
+}
 export type ConvertibleToDisposable = ((() => void) | IDisposable | Disposable | null | undefined);
 
 class NamedDisposable implements IDisposable {