+++ browser/mainapp/src/viewmodel/sidebartabs/ChatsListTabViewModel.ts
@@ -1,0 +1,10 @@
+import { ActiveLoginViewModel } from "../ActiveLoginViewModel";
+import { StandardSidebarTabViewModel } from "./StandardSidebarTabViewModel";
+
+export class ChatsListTabViewModel extends StandardSidebarTabViewModel {
+    constructor(public readonly session: ActiveLoginViewModel) {
+        super();
+    }
+
+    readonly tabId = "chatslist";
+}